add ouq_misc, ytrue-directed surrogate examples
